React: refactor duplicate code, remove SlideUpTransition (#102)
Move duplicate code into utils.tsx
