+++ react/src/pages/Notification.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { makeStyles, Tooltip, Button } from '@material-ui/core';
 import { Alert } from '@material-ui/lab';
-import { PipelineStatus, AnalysisRow } from './analysis/Analysis';
+import { PipelineStatus, Analysis } from './utils';
 
 const useStyles = makeStyles(theme => ({
     msgBox: {
@@ -27,7 +27,7 @@ const useStyles = makeStyles(theme => ({
 }));
 
 export interface NotificationProps {
-    analysis: AnalysisRow,
+    analysis: Analysis,
 }
 
 enum Severity {
@@ -50,7 +50,7 @@ function createNotification(
     return { analysis_id, msg, severity };
 }
 
-const getNotificationInfo = (analysis: AnalysisRow) => {
+const getNotificationInfo = (analysis: Analysis) => {
     const msg = `The status of ${analysis.analysis_id} has changed to `;
     switch (analysis.state) {
         case PipelineStatus.PENDING:
