+++ react/src/pages/analysis/DatasetTable.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import { Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Paper } from '@material-ui/core';
 
-
 const useStyles = makeStyles({
     table: {
         minWidth: 650,
