--- react/src/pages/dashboard/MockData.tsx
@@ -1,70 +1,0 @@
-export interface Participant {
-    participant_id: string,
-    participant_codename: string,
-    family_id: string,
-    family_codename: string,
-    participant_type: string,
-    affected: boolean,
-    solved: boolean,
-    sex: string,
-    notes: string,
-    dataset_types: string[],
-    created: string,
-    created_by: number,
-    updated: string,
-    updated_by: number,
-    tissue_samples: Sample[]
-}
-
-export interface Sample {
-    sampleID: string,
-    extractionDate: string,
-    sampleType: string,
-    tissueProcessing: string,
-    datasets: Dataset[],
-    notes: string,
-    created: string,
-    createBy: number,
-    updated: string,
-    updatedBy: number,
-}
-
-export interface Dataset {
-    dataset_id: string,
-    tissue_sample_id: string,
-    dataset_type: string,
-    input_hpf_path: string,
-    notes: string,
-    condition: string,
-    extraction_protocol: string,
-    capture_kit: string,
-    library_prep_method: string,
-    library_prep_date: string,
-    read_length: number,
-    read_type: string,
-    sequencing_id: string,
-    sequencing_centre: string,
-    batch_id: string,
-    created: string,
-    created_by: number,
-    updated: string,
-    updated_by: number,
-    discriminator: string,
-}
-
-export interface Analysis {
-    analysisID: string,
-    datasetID: string,
-    analysisState: string,
-    pipelineID: string,
-    qsubID: string,
-    resultHpfPath: string,
-    assignee: number,
-    requester: number,
-    requested: string,
-    started: string,
-    finished: string,
-    notes: string,
-    updated: string,
-    updatedBy: number,
-}
