+++ react/src/pages/utils.tsx
@@ -1,16 +1,115 @@
-import React, { forwardRef, ReactElement, Ref, useState } from "react";
-import { IconButton, Slide, Snackbar } from "@material-ui/core";
-import { TransitionProps } from "@material-ui/core/transitions";
-import { Close } from "@material-ui/icons";
+/*****   CONSTANTS   *****/
+export const emptyCellValue = "<empty>";
 
-export const SlideUpTransition = forwardRef((
-    props: TransitionProps & { children?: ReactElement },
-    ref: Ref<unknown>,
-) => <Slide direction="up" ref={ref} {...props} />);
+const months = [
+    "Jan",
+    "Feb",
+    "Mar",
+    "Apr",
+    "May",
+    "Jun",
+    "Jul",
+    "Aug",
+    "Sep",
+    "Oct",
+    "Nov",
+    "Dec"
+];
 
+/*****   TYPINGS   *****/
 export type Counts = { [key: string]: number };
 export type KeyValue = { [key: string]: string };
 
+/*****   ENUMS   *****/
+export enum PipelineStatus {
+    PENDING = "Pending",
+    RUNNING = "Running",
+    COMPLETED = "Completed",
+    ERROR = "Error",
+    CANCELLED = "Cancelled"
+}
+
+/*****   INTERFACES   *****/
+export interface Participant {
+    participant_id: string,
+    participant_codename: string,
+    family_id: string,
+    family_codename: string,
+    participant_type: string,
+    affected: boolean,
+    solved: boolean,
+    sex: string,
+    notes: string,
+    dataset_types: string[],
+    created: string,
+    created_by: number,
+    updated: string,
+    updated_by: number,
+    tissue_samples: Sample[]
+}
+export interface Sample {
+    sampleID: string,
+    extractionDate: string,
+    sampleType: string,
+    tissueProcessing: string,
+    datasets: Dataset[],
+    notes: string,
+    created: string,
+    createBy: number,
+    updated: string,
+    updatedBy: number,
+}
+export interface Dataset {
+    dataset_id: string,
+    participant_codename: string,
+    family_codename: string,
+    tissue_sample_type: string,
+    tissue_sample_id: string,
+    dataset_type: string,
+    input_hpf_path: string,
+    notes: string,
+    condition: string,
+    extraction_protocol: string,
+    capture_kit: string,
+    library_prep_method: string,
+    library_prep_date: string,
+    read_length: number,
+    read_type: string,
+    sequencing_id: string,
+    sequencing_centre: string,
+    batch_id: string,
+    created: string,
+    created_by: number,
+    updated: string,
+    updated_by: number,
+    discriminator: string,
+}
+export interface Analysis {
+    analysis_id: string,
+    pipeline_id: string,
+    result_hpf_path: string,
+    assignee: string,
+    requester: string,
+    state: PipelineStatus,
+    updated: string,
+    notes: string,
+    selected: boolean,
+    datasetID: string,
+    analysisState: string,
+    qsubID: string,
+    requested: string,
+    started: string,
+    finished: string,
+    updatedBy: number,
+}
+export interface Pipeline {
+    pipeline_id: number;
+    pipeline_name: string;
+    pipeline_version: string;
+    supported_types: string[];
+}
+
+/*****   FUNCTIONS   *****/
 export function countArray(items: string[]) {
     return items.reduce<Counts>((counts, item) => {
         if (counts[item]) {
@@ -28,21 +127,6 @@ export function toKeyValue(items: string[]) {
         return map;
     }, Object.create(null));
 }
-
-const months = [
-    "Jan",
-    "Feb",
-    "Mar",
-    "Apr",
-    "May",
-    "Jun",
-    "Jul",
-    "Aug",
-    "Sep",
-    "Oct",
-    "Nov",
-    "Dec"
-];
 
 /**
  * Return a date string in the format "YYYY-MM-DD", if possible.
@@ -62,4 +146,32 @@ export function formatDateString(date: string) {
     return date;
 }
 
-export const emptyCellValue = "<empty>";
+/**
+ * Convert the provided JSON Array to a valid array of Analysis.
+ */
+export function jsonToAnalyses(data: Array<any>): Analysis[] {
+    const rows: Analysis[] = data.map((row, index, arr) => {
+        switch (row.analysis_state) {
+            case 'Requested':
+                row.state = PipelineStatus.PENDING;
+                break;
+            case 'Running':
+                row.state = PipelineStatus.RUNNING;
+                break;
+            case 'Done':
+                row.state = PipelineStatus.COMPLETED;
+                break;
+            case 'Error':
+                row.state = PipelineStatus.ERROR;
+                break;
+            case 'Cancelled':
+                row.state = PipelineStatus.CANCELLED;
+                break;
+            default:
+                row.state = null;
+                break;
+        }
+        return { ...row, selected: false } as Analysis;
+    });
+    return rows;
+}
